requirements.txt
apache-beam[gcp]
google-cloud-bigquery
google-cloud-storage
google-cloud-secret-manager
requests
python-dotenv